begun breaking board.rs into separate submodules
